Cut the truncated message
